+++ src/task/confirmPayTransaction.ts
@@ -8,6 +8,10 @@ export type IPurpose = IConfirmPayPurpose & {
 };
 export interface IData extends IConfirmPayActionAttributes {
     purpose: IPurpose;
+    /**
+     * OrderPaymentDue->OrderProcessing処理を実行するかどうかのオプションを追加(2023-08-23~)
+     */
+    processOrder?: boolean;
 }
 export interface IAttributes extends TaskFactory.IAttributes {
     name: TaskName.ConfirmPayTransaction;