+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 
+- 注文のpaymentMethodsにpaymentStatusを追加
+- 資産取引検索条件拡張
+- 決済取引確定タスクにprocessOrderオプションを追加
+
 ### Changed
 
 ### Deprecated